+++ apps/heft/src/pluginFramework/logging/LoggingManager.ts
@@ -2,7 +2,12 @@
 // See LICENSE in the project root for license information.
 
 import { ScopedLogger } from './ScopedLogger';
-import { ITerminalProvider, FileError, FileErrorFormat } from '@rushstack/node-core-library';
+import {
+  FileError,
+  FileLocationStyle,
+  ITerminalProvider,
+  IFileErrorFormattingOptions
+} from '@rushstack/node-core-library';
 
 export interface ILoggingManagerOptions {
   terminalProvider: ITerminalProvider;
@@ -42,13 +47,15 @@ export class LoggingManager {
     }
   }
 
-  public getErrorStrings(fileErrorFormat?: FileErrorFormat): string[] {
+  public getErrorStrings(fileLocationStyle?: FileLocationStyle): string[] {
     const result: string[] = [];
 
     for (const scopedLogger of this._scopedLoggers.values()) {
       result.push(
         ...scopedLogger.errors.map(
-          (error) => `[${scopedLogger.loggerName}] ${LoggingManager.getErrorMessage(error, fileErrorFormat)}`
+          (error) =>
+            `[${scopedLogger.loggerName}] ` +
+            LoggingManager.getErrorMessage(error, { format: fileLocationStyle })
         )
       );
     }
@@ -56,14 +63,15 @@ export class LoggingManager {
     return result;
   }
 
-  public getWarningStrings(fileErrorFormat?: FileErrorFormat): string[] {
+  public getWarningStrings(fileErrorFormat?: FileLocationStyle): string[] {
     const result: string[] = [];
 
     for (const scopedLogger of this._scopedLoggers.values()) {
       result.push(
         ...scopedLogger.warnings.map(
           (warning) =>
-            `[${scopedLogger.loggerName}] ${LoggingManager.getErrorMessage(warning, fileErrorFormat)}`
+            `[${scopedLogger.loggerName}] ` +
+            LoggingManager.getErrorMessage(warning, { format: fileErrorFormat })
         )
       );
     }
@@ -71,9 +79,9 @@ export class LoggingManager {
     return result;
   }
 
-  public static getErrorMessage(error: Error, fileErrorFormat?: FileErrorFormat): string {
+  public static getErrorMessage(error: Error, options?: IFileErrorFormattingOptions): string {
     if (error instanceof FileError) {
-      return error.toString(fileErrorFormat);
+      return error.getFormattedErrorMessage(options);
     } else {
       return error.message;
     }