+++ heft-plugins/heft-typescript-plugin/src/TypeScriptBuilder.ts
@@ -539,12 +539,13 @@ export class TypeScriptBuilder {
     if (diagnostic.file) {
       const { line, character } = diagnostic.file.getLineAndCharacterOfPosition(diagnostic.start!);
       const message: string = ts.flattenDiagnosticMessageText(diagnostic.messageText, '\n');
-      const buildFolderRelativeFilename: string = path.relative(
-        this._configuration.buildFolder,
-        diagnostic.file.fileName
-      );
       const formattedMessage: string = `(TS${diagnostic.code}) ${message}`;
-      errorObject = new FileError(formattedMessage, buildFolderRelativeFilename, line + 1, character + 1);
+      errorObject = new FileError(formattedMessage, {
+        absolutePath: diagnostic.file.fileName,
+        projectFolder: this._configuration.buildFolder,
+        line: line + 1,
+        column: character + 1
+      });
       diagnosticMessage = errorObject.toString();
     } else {
       diagnosticMessage = ts.flattenDiagnosticMessageText(diagnostic.messageText, '\n');