+++ libraries/debug-certificate-manager/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log - @rushstack/debug-certificate-manager
 
-This log was last generated on Fri, 24 Jun 2022 07:16:47 GMT and should not be manually modified.
+This log was last generated on Sat, 25 Jun 2022 21:00:40 GMT and should not be manually modified.
+
+## 1.1.53
+Sat, 25 Jun 2022 21:00:40 GMT
+
+_Version update only_
+
+## 1.1.52
+Sat, 25 Jun 2022 01:54:29 GMT
+
+_Version update only_
 
 ## 1.1.51
 Fri, 24 Jun 2022 07:16:47 GMT